requirements.txt
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.24.0
pandas-ta>=0.3.14b
requests>=2.28.0
supabase>=2.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
jsonschema>=4.22.0
